change basic type f64 to f32

+++ src/camera.rs
@@ -8,7 +8,7 @@ pub struct Camera {
     vertical: Vec3,
     cu: Vec3,
     cv: Vec3,
-    lens_radius: f64,
+    lens_radius: f32,
 }
 
 impl Camera {
@@ -16,13 +16,13 @@ impl Camera {
         lookfrom: Point3,
         lookat: Point3,
         vup: Vec3,
-        vfov: f64,
-        aspect_ratio: f64,
-        aperture: f64,
-        focus_dist: f64,
+        vfov: f32,
+        aspect_ratio: f32,
+        aperture: f32,
+        focus_dist: f32,
     ) -> Camera {
         // Vertical field-of-view in degrees
-        let theta = std::f64::consts::PI / 180.0 * vfov;
+        let theta = std::f32::consts::PI / 180.0 * vfov;
         let viewport_height = 2.0 * (theta / 2.0).tan();
         let viewport_width = aspect_ratio * viewport_height;
 
@@ -45,7 +45,7 @@ impl Camera {
         }
     }
 
-    pub fn get_ray(&self, s: f64, t: f64) -> Ray {
+    pub fn get_ray(&self, s: f32, t: f32) -> Ray {
         let rd = self.lens_radius * Vec3::random_in_unit_disk();
         let offset = rd.x() * self.cu + rd.y() * self.cv;
         Ray::new(
