+++ src/material.rs
@@ -35,11 +35,11 @@ impl Scatter for Lambertian {
 
 pub struct Metal {
     albedo: Color,
-    fuzz: f64,
+    fuzz: f32,
 }
 
 impl Metal {
-    pub fn new(a: Color, f: f64) -> Metal {
+    pub fn new(a: Color, f: f32) -> Metal {
         Metal { albedo: a, fuzz: f }
     }
 }
@@ -58,17 +58,17 @@ impl Scatter for Metal {
 }
 
 pub struct Dielectric {
-    ir: f64,
+    ir: f32,
 }
 
 impl Dielectric {
-    pub fn new(index_of_refraction: f64) -> Dielectric {
+    pub fn new(index_of_refraction: f32) -> Dielectric {
         Dielectric {
             ir: index_of_refraction,
         }
     }
 
-    fn reflectance(cosine: f64, ref_idx: f64) -> f64 {
+    fn reflectance(cosine: f32, ref_idx: f32) -> f32 {
         // Use Schlick's approximation for reflectance
         let r0 = ((1.0 - ref_idx) / (1.0 + ref_idx)).powi(2);
         r0 + (1.0 - r0) * (1.0 - cosine).powi(5)
@@ -89,7 +89,7 @@ impl Scatter for Dielectric {
 
         let mut rng = rand::thread_rng();
         let cannot_refract = refraction_ratio * sin_theta > 1.0;
-        let will_reflect = rng.gen::<f64>() < Self::reflectance(cos_theta, refraction_ratio);
+        let will_reflect = rng.gen::<f32>() < Self::reflectance(cos_theta, refraction_ratio);
 
         let direction = if cannot_refract || will_reflect {
             unit_dir.reflect(rec.normal)
